requirements.txt
Flask
python-dotenv
joblib
scikit-learn 